Implement BartXmlTask abstract class
