+++ Android/mobile/src/main/java/edu/berkeley/eecs/bartgo/BartXmlTask.java
@@ -1,0 +1,49 @@
+package edu.berkeley.eecs.bartgo;
+
+import android.os.AsyncTask;
+import android.util.Log;
+
+import org.xmlpull.v1.XmlPullParserException;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.net.URL;
+import java.net.HttpURLConnection;
+
+abstract class BartXmlTask extends AsyncTask<String, Void, String> {
+    private final String TAG = "BartXmlTask";
+
+    @Override
+    protected String doInBackground(String... urls) {
+        try {
+            return loadXmlFromNetwork(urls[0]);
+        } catch (IOException e) {
+            Log.e(TAG, String.valueOf(e));
+            return "Connection error";
+        } catch (XmlPullParserException e) {
+            Log.e(TAG, String.valueOf(e));
+            return "XML Error:" + e;
+        }
+    }
+
+    @Override
+    protected void onPostExecute(String result) {
+        super.onPostExecute(result);
+    }
+
+    abstract String loadXmlFromNetwork(String urlString) throws XmlPullParserException, IOException;
+
+    // Given a string representation of a URL, sets up a connection and gets
+    // an input stream.
+    protected InputStream downloadUrl(String urlString) throws IOException {
+        URL url = new URL(urlString);
+        HttpURLConnection conn = (HttpURLConnection) url.openConnection();
+        conn.setReadTimeout(10000 /* milliseconds */);
+        conn.setConnectTimeout(15000 /* milliseconds */);
+        conn.setRequestMethod("GET");
+        conn.setDoInput(true);
+        // Starts the query
+        conn.connect();
+        return conn.getInputStream();
+    }
+}
